prototype/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1
markdown==3.10.2
sse-starlette==2.2.1
pytest==8.3.4
pytest-asyncio==0.25.0
httpx[http2]==0.28.1
behave==1.2.6
